+++ ghost-compiler/src/main/kotlin/com/ghost/serialization/compiler/FlattenOptionsGenerator.kt
@@ -102,11 +102,30 @@ internal object FlattenOptionsGenerator {
             }
         }.distinct()
 
-        val (shift, multiplier, tableSize) = PerfectHashFinder.findPerfectHash(names)
+        val hashConfig = PerfectHashFinder.findPerfectHash(names)
 
         val optionsClass = readerClass.peerClass(C.STR_OPTIONS_CLASS)
         val optionsBuilder = CodeBlock.builder()
-            .add(C.TEMPLATE_OPTIONS_OF_SEEDS_START, optionsClass, shift, multiplier, tableSize, textChannel)
+        if (hashConfig.extendedKeyHash) {
+            optionsBuilder.add(
+                C.TEMPLATE_OPTIONS_OF_SEEDS_EXTENDED_START,
+                optionsClass,
+                hashConfig.shift,
+                hashConfig.multiplier,
+                hashConfig.tableSize,
+                textChannel,
+                true
+            )
+        } else {
+            optionsBuilder.add(
+                C.TEMPLATE_OPTIONS_OF_SEEDS_START,
+                optionsClass,
+                hashConfig.shift,
+                hashConfig.multiplier,
+                hashConfig.tableSize,
+                textChannel
+            )
+        }
 
         names.forEach { name ->
             optionsBuilder.add(C.TEMPLATE_COMMA_FORMAT_S, name)